Base migration to support new mv5 core
