--- src/main/java/com/onarandombox/multiverseinventories/command/package-info.java
@@ -1,5 +1,0 @@
-/**
- * This package contains all Commands.
- */
-package com.onarandombox.multiverseinventories.command;
-
