+++ src/main/java/com/onarandombox/multiverseinventories/commands/package-info.java
@@ -1,0 +1,5 @@
+/**
+ * This package contains all Commands.
+ */
+package com.onarandombox.multiverseinventories.commands;
+
